# swagger/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Digitaler Zwilling Simulations-API",
    "description": "Eine API zur Durchführung von Simulationen eines digitalen Zwillings in Factory I/O, welche via eines WfMS gestartet und kontrolliert werden."
  },
  "paths": {
    "/tasks/details": {
      "post": {
        "summary": "Ruft Details anhand der Referenz ab",
        "tags": [
          "Informationen abrufen"
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/Reference"
        },
        "responses": {
          "201": {
            "description": "Ok"
          }
        }
      }
    },
...
    "/tasks/varnishing": {
      "put": {
        "summary": "Startet die Lackierung des Produktes",
        "parameters": [
          {
            "in": "header",
            "name": "cpee-callback",
            "schema": {
              "type": "string",
              "example": "yolomadonna"
            },
            "required": true
          },
          {
            "in": "header",
            "name": "cpee-instance",
            "schema": {
              "type": "integer",
              "example": 100
            },
            "required": true
          },
          {
            "in": "header",
            "name": "cpee-label",
            "schema": {
              "type": "string",
              "example": "Oberflaechenbehandlung"
            },
            "required": true
          }
        ],
        "tags": [
          "Tasks (Beide)"
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/Reference"
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/tasks/addToPallet": {
      "put": {
        "summary": "Hinterlegt Referenz auf RFID-Transponder der Palette, auf welche das Paket gestapelt wird",
        "parameters": [
          {
            "in": "header",
            "name": "cpee-callback",
            "schema": {
              "type": "string",
              "example": "yolomadonna"
            },
            "required": true
          },
          {
            "in": "header",
            "name": "cpee-instance",
            "schema": {
              "type": "integer",
              "example": 100
            },
            "required": true
          },
          {
            "in": "header",
            "name": "cpee-label",
            "schema": {
              "type": "string",
              "example": "Paket auf Palette beladen und abwarten bis diese voll ist"
            },
            "required": true
          }
        ],
        "tags": [
          "Tasks (Beide)"
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/Reference"
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/tasks/saveSchrank": {
      "post": {
        "summary": "Hinterlegt Referenznummer in Liste",
        "tags": [
          "Referenznummer sichern oder abrufen"
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/Reference"
        },
        "responses": {
          "201": {
            "description": "Added"
          }
        }
      }
    },
    "/tasks/saveSchreibtisch": {
      "post": {
        "summary": "Hinterlegt Referenznummer in Liste",
        "tags": [
          "Referenznummer sichern oder abrufen"
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/Reference"
        },
        "responses": {
          "201": {
            "description": "Added"
          }
        }
      }
    },
    "/tasks/mirrorReference": {
      "post": {
        "summary": "Hinterlegt Referenznummer in Liste",
        "tags": [
          "Referenznummer sichern oder abrufen"
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/Reference"
        },
        "responses": {
          "201": {
            "description": "Added"
          }
        }
      }
    },
    "/tasks/boxReference": {
      "post": {
        "summary": "Hinterlegt Referenznummer in Liste",
        "tags": [
          "Referenznummer sichern oder abrufen"
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/Reference"
        },
        "responses": {
          "201": {
            "description": "Added"
          }
        }
      }
    },
    "/tasks/init": {
      "get": {
        "summary": "Initialisiert die Simulation und schreibt Werte auf den RFID Transponder",
        "parameters": [
          {
            "in": "header",
            "name": "cpee-callback",
            "schema": {
              "type": "string",
              "example": "yolomadonna"
            },
            "required": true
          },
          {
            "in": "header",
            "name": "cpee-label",
            "schema": {
              "type": "string",
              "example": "RFID Transponder initialisieren"
            },
            "required": true
          }
        ],
        "tags": [
          "Tasks (Beide)"
        ],
        "responses": {
          "200": {
            "description": "Ok"
          }
        }
      }
    },
    "/tasks/start": {
      "post": {
        "parameters": [
          {
            "in": "header",
            "name": "cpee-callback",
            "schema": {
              "type": "string",
              "example": "yolomadonna"
            },
            "required": true
          },
          {
            "in": "header",
            "name": "cpee-instance",
            "schema": {
              "type": "integer",
              "example": 100
            },
            "required": true
          },
          {
            "in": "header",
            "name": "cpee-label",
            "schema": {
              "type": "string",
              "example": "Material auf Palette vorbereiten"
            },
            "required": true
          }
        ],
        "summary": "Leitet eine Simulation in Factory I/O ein.",
        "tags": [
          "Tasks (Beide)"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/Schrank"
                  },
                  {
                    "$ref": "#/components/schemas/Schreibtisch"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/tasks/machining": {
      "put": {
        "summary": "Startet die erforderliche Maschine",
        "parameters": [
          {
            "in": "header",
            "name": "cpee-callback",
            "schema": {
              "type": "string",
              "example": "yolomadonna"
            },
            "required": true
          },
          {
            "in": "header",
            "name": "cpee-instance",
            "schema": {
              "type": "integer",
              "example": 100
            },
            "required": true
          },
          {
            "in": "header",
            "name": "cpee-label",
            "schema": {
              "type": "string",
              "example": "Material zuschneiden mit Maschine A"
            },
            "required": true
          }
        ],
        "tags": [
          "Tasks (Beide)"
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/Machine"
        },
        "responses": {
          "201": {
            "description": "Updated"
          }
        }
      }
    },
    "/tasks/height": {
      "put": {
        "summary": "Startet das Abmessen des Paketes",
        "parameters": [
          {
            "in": "header",
            "name": "cpee-callback",
            "schema": {
              "type": "string",
              "example": "yolomadonna"
            },
            "required": true
          },
          {
            "in": "header",
            "name": "cpee-instance",
            "schema": {
              "type": "integer",
              "example": 100
            },
            "required": true
          },
          {
            "in": "header",
            "name": "cpee-label",
            "schema": {
              "type": "string",
              "example": "Pakethoehe messen"
            },
            "required": true
          }
        ],
        "tags": [
          "Tasks (Beide)"
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/Reference"
        },
        "responses": {
          "201": {
            "description": "Ok"
          }
        }
      }
    },
    "/tasks/weight": {
      "put": {
        "summary": "Startet das Abwiegen des Paketes",
        "parameters": [
          {
            "in": "header",
            "name": "cpee-callback",
            "schema": {
              "type": "string",
              "example": "yolomadonna"
            },
            "required": true
          },
          {
            "in": "header",
            "name": "cpee-instance",
            "schema": {
              "type": "integer",
              "example": 100
            },
            "required": true
          },
          {
            "in": "header",
            "name": "cpee-label",
            "schema": {
              "type": "string",
              "example": "Paket wiegen"
            },
            "required": true
          }
        ],
        "tags": [
          "Tasks (Beide)"
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/Reference"
        },
        "responses": {
          "201": {
            "description": "Ok"
          }
        }
      }
    },
    "/tasks/preDrill": {
      "put": {
        "summary": "Startet das Vorbohren des Materials",
        "parameters": [
          {
            "in": "header",
            "name": "cpee-callback",
            "schema": {
              "type": "string",
              "example": "yolomadonna"
            },
            "required": true
          },
          {
            "in": "header",
            "name": "cpee-instance",
            "schema": {
              "type": "integer",
              "example": 100
            },
            "required": true
          },
          {
            "in": "header",
            "name": "cpee-label",
            "schema": {
              "type": "string",
              "example": "Loecher vorbohren (Schreibtisch)"
            },
            "required": true
          }
        ],
        "tags": [
          "Tasks (Beide)"
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/Reference"
        },
        "responses": {
          "201": {
            "description": "Ok"
          }
        }
      }
    },
    "/tasks/drawers": {
      "put": {
        "summary": "Startet die Herstellung der Schubladen",
        "parameters": [
          {
            "in": "header",
            "name": "cpee-callback",
            "schema": {
              "type": "string",
              "example": "yolomadonna"
            },
            "required": true
          },
          {
            "in": "header",
            "name": "cpee-instance",
            "schema": {
              "type": "integer",
              "example": 100
            },
            "required": true
          },
          {
            "in": "header",
            "name": "cpee-label",
            "schema": {
              "type": "string",
              "example": "Schubladen herstellen"
            },
            "required": true
          }
        ],
        "tags": [
          "Tasks (Schreibtisch)"
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/Reference"
        },
        "responses": {
          "201": {
            "description": "Ok"
          }
        }
      }
    },
    "/tasks/shelves": {
      "put": {
        "summary": "Startet die Herstellung der Regale",
        "parameters": [
          {
            "in": "header",
            "name": "cpee-callback",
            "schema": {
              "type": "string",
              "example": "yolomadonna"
            },
            "required": true
          },
          {
            "in": "header",
            "name": "cpee-instance",
            "schema": {
              "type": "integer",
              "example": 100
            },
            "required": true
          },
          {
            "in": "header",
            "name": "cpee-label",
            "schema": {
              "type": "string",
              "example": "Regale herstellen"
            },
            "required": true
          }
        ],
        "tags": [
          "Tasks (Schrank)"
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/Reference"
        },
        "responses": {
          "201": {
            "description": "Ok"
          }
        }
      }
    },
    "/tasks/defaultDoor": {
      "put": {
        "summary": "Startet die Herstellung der Default-Tür",
        "parameters": [
          {
            "in": "header",
            "name": "cpee-callback",
            "schema": {
              "type": "string",
              "example": "yolomadonna"
            },
            "required": true
          },
          {
            "in": "header",
            "name": "cpee-instance",
            "schema": {
              "type": "integer",
              "example": 100
            },
            "required": true
          },
          {
            "in": "header",
            "name": "cpee-label",
            "schema": {
              "type": "string",
              "example": "Standardtuer herstellen"
            },
            "required": true
          }
        ],
        "tags": [
          "Tasks (Schrank)"
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/Reference"
        },
        "responses": {
          "201": {
            "description": "Ok"
          }
        }
      }
    },
    "/tasks/slidingDoor": {
      "put": {
        "summary": "Startet die Herstellung der Schiebetür",
        "parameters": [
          {
            "in": "header",
            "name": "cpee-callback",
            "schema": {
              "type": "string",
              "example": "yolomadonna"
            },
            "required": true
          },
          {
            "in": "header",
            "name": "cpee-instance",
            "schema": {
              "type": "integer",
              "example": 100
            },
            "required": true
          },
          {
            "in": "header",
            "name": "cpee-label",
            "schema": {
              "type": "string",
              "example": "Schiebetuer herstellen"
            },
            "required": true
          }
        ],
        "tags": [
          "Tasks (Schrank)"
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/Reference"
        },
        "responses": {
          "201": {
            "description": "Ok"
          }
        }
      }
    },
    "/tasks/lock": {
      "put": {
        "summary": "Startet das Einbauen eines Türschlosses",
        "parameters": [
          {
            "in": "header",
            "name": "cpee-callback",
            "schema": {
              "type": "string",
              "example": "yolomadonna"
            },
            "required": true
          },
          {
            "in": "header",
            "name": "cpee-instance",
            "schema": {
              "type": "integer",
              "example": 100
            },
            "required": true
          },
          {
            "in": "header",
            "name": "cpee-label",
            "schema": {
              "type": "string",
              "example": "Tuerschloss einbauen"
            },
            "required": true
          }
        ],
        "tags": [
          "Tasks (Schrank)"
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/Reference"
        },
        "responses": {
          "201": {
            "description": "Ok"
          }
        }
      }
    },
...
  },
  "components": {
    "requestBodies": {
...
      "Reference": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Reference"
            }
          }
        }
      },
...
    },
    "schemas": {
      "Schrank": {
        "type": "object",
        "properties": {
          "product": {
            "type": "string",
            "example": "schrank"
          },
          "mirrorShape": {
            "type": "string",
            "example": "rund"
          },
          "doorType": {
            "type": "string",
            "example": "default"
          },
          "express": {
            "type": "boolean",
            "example": true
          },
          "additionalEquipment": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "Schreibtisch": {
        "type": "object",
        "properties": {
          "product": {
            "type": "string",
            "example": "schreibtisch"
          },
          "tableStyle": {
            "type": "string",
            "example": "classic"
          },
          "express": {
            "type": "boolean",
            "example": true
          },
          "additionalEquipment": {
            "type": "boolean",
            "example": true
          }
        }
      },
...
    }
  },
  "servers": [
    {
      "description": "Simulations-API",
      "url": "https://localhost:9033"
    }
  ]
}